Df/fixes 224 number logic (#263)
* Fixes #224 (Clarify rounding Logic in NumberScalingUtils)
* Fixes #248 (Make SettlementQuantity.amount into a BigInteger)
* Misc cleanup
* Remove unused cache
* Improve unit test coverage for SettlementQuantity

+++ ilpv4-connector-service-impl/src/main/java/com/sappenin/interledger/ilpv4/connector/links/filters/OutgoingBalanceLinkFilter.java
@@ -114,7 +114,7 @@ public class OutgoingBalanceLinkFilter extends AbstractLinkFilter implements Lin
                   final UUID idempotencyId = UUID.randomUUID();
 
                   final SettlementQuantity settlementQuantityInClearingUnits = SettlementQuantity.builder()
-                    .amount(balanceForFulfillResponse.clearingAmountToSettle())
+                    .amount(BigInteger.valueOf(balanceForFulfillResponse.clearingAmountToSettle()))
                     .scale(destinationAccountSettings.getAssetScale())
                     .build();
